ft/abc: Move to modular design and clarify variables
As I mentioned in 33dbac2, I've been considering moving to a modular
design that would reduce the number of abc#getNotes() and
abc#getNotation() calls when repeat or stringing together functions. As
all planned functions work on the note list rather than raw abc, I've
decided to move to that modular design.
This means it is now a pain to run these functions normally (what used
to be just `abc#lengthDouble(abcInput)` now requires
`abc#getNotation(abcInput, abc#lengthDouble(abc#getNotes(abcInput)))`.
Obviously this is stupid, so I've created the abc#run() function that
does this for you. It takes the original abcInput and the function name
(using the same example as before, `abc#lengthDouble`) as inputs.

The variables have been clarified (from just using 'input' everywhere)
to 'abcInput' for every function that deals with raw input and 'notelist'
for functions that just deal with the 2D Arrays.